post long message intent confirmation
