+++ lambda/custom/handlers/General/NoIntentHandler.js
@@ -1,8 +1,4 @@
 const { ri } = require('@jargon/alexa-skill-sdk');
-const { login, postMessage } = require('../../helperFunctions');
-const { supportsAPL } = require('../../utils');
-const titleMessageBoxTemplate = require('../../APL/templates/titleMessageBoxTemplate');
-
 
 const NoIntentHandler = {
 	canHandle(handlerInput) {
@@ -15,44 +11,21 @@ const NoIntentHandler = {
 			const sessionAttributes = attributesManager.getSessionAttributes() || {};
 
 			if (sessionAttributes.postLongMessageIntentOnProgress) {
-				delete sessionAttributes.postLongMessageIntentOnProgress;
-				delete sessionAttributes.channelConfirm;
-				const {
-					accessToken,
-				} = handlerInput.requestEnvelope.context.System.user;
 
-
-
-				const { channelName } = sessionAttributes;
-				const { message } = sessionAttributes;
-
-				delete sessionAttributes.channelName;
-				delete sessionAttributes.message;
-
-				const headers = await login(accessToken);
-				const speechText = await postMessage(channelName, message, headers);
-				const repromptText = ri('GENERIC_REPROMPT');
-
-
-				if (supportsAPL(handlerInput)) {
-					const data = {
-						title: handlerInput.translate('POST_MESSAGE.APL_SUCCESS', { channelName }),
-						message,
-					};
-
+				if (!sessionAttributes.postLongMessageConfirmation) {
+					sessionAttributes.postLongMessageConfirmation = true;
+					const speechText = ri('POST_MESSAGE.POST_LONG_MESSAGE_CONFIRMATION', { longmessage: sessionAttributes.message, channelName: sessionAttributes.channelName });
 					return handlerInput.jrb
 						.speak(speechText)
-						.speak(repromptText)
-						.reprompt(repromptText)
-						.addDirective(titleMessageBoxTemplate(data))
+						.reprompt(speechText)
 						.getResponse();
-
 				} else {
+					const speechText = ri('POST_MESSAGE.DENIED');
+					const repromptText = ri('GENERIC_REPROMPT');
 					return handlerInput.jrb
 						.speak(speechText)
 						.speak(repromptText)
 						.reprompt(repromptText)
-						.withSimpleCard(ri('POST_MESSAGE.CARD_TITLE'), speechText)
 						.getResponse();
 				}
 			} else {
