+++ lambda/custom/handlers/General/YesIntentHandler.js
@@ -1,34 +1,86 @@
 const { ri } = require('@jargon/alexa-skill-sdk');
+const { login, postMessage } = require('../../helperFunctions');
+const { supportsAPL } = require('../../utils');
+const titleMessageBoxTemplate = require('../../APL/templates/titleMessageBoxTemplate');
 
 const YesIntentHandler = {
 	canHandle(handlerInput) {
 		return handlerInput.requestEnvelope.request.type === 'IntentRequest' &&
 			handlerInput.requestEnvelope.request.intent.name === 'AMAZON.YesIntent';
 	},
-	handle(handlerInput) {
-
+	async handle(handlerInput) {
 		const { attributesManager } = handlerInput;
 		const sessionAttributes = attributesManager.getSessionAttributes() || {};
 
-		return handlerInput.jrb
-			.speak(ri('POST_MESSAGE.CONFIRM_MORE'))
-			.reprompt(ri('POST_MESSAGE.CONFIRM_MORE_REPROMPT'))
-			.addElicitSlotDirective('longmessage', {
-				name: 'PostLongMessageIntent',
-				confirmationStatus: 'NONE',
-				slots: {
-					channelname: {
-						name: 'channelname',
-						value: sessionAttributes.channelName,
-						confirmationStatus: 'NONE',
+		if (sessionAttributes.postLongMessageIntentOnProgress) {
+
+			if (sessionAttributes.postLongMessageConfirmation) {
+
+				delete sessionAttributes.postLongMessageConfirmation;
+				delete sessionAttributes.postLongMessageIntentOnProgress;
+				delete sessionAttributes.channelConfirm;
+				const {
+					accessToken,
+				} = handlerInput.requestEnvelope.context.System.user;
+
+
+
+				const { channelName } = sessionAttributes;
+				const { message } = sessionAttributes;
+
+				delete sessionAttributes.channelName;
+				delete sessionAttributes.message;
+
+				const headers = await login(accessToken);
+				const speechText = await postMessage(channelName, message, headers);
+				const repromptText = ri('GENERIC_REPROMPT');
+
+
+				if (supportsAPL(handlerInput)) {
+					const data = {
+						title: handlerInput.translate('POST_MESSAGE.APL_SUCCESS', { channelName }),
+						message,
+					};
+
+					return handlerInput.jrb
+						.speak(speechText)
+						.speak(repromptText)
+						.reprompt(repromptText)
+						.addDirective(titleMessageBoxTemplate(data))
+						.getResponse();
+
+				} else {
+					return handlerInput.jrb
+						.speak(speechText)
+						.speak(repromptText)
+						.reprompt(repromptText)
+						.withSimpleCard(ri('POST_MESSAGE.CARD_TITLE'), speechText)
+						.getResponse();
+				}
+			}
+
+			return handlerInput.jrb
+				.speak(ri('POST_MESSAGE.CONFIRM_MORE'))
+				.reprompt(ri('POST_MESSAGE.CONFIRM_MORE_REPROMPT'))
+				.addElicitSlotDirective('longmessage', {
+					name: 'PostLongMessageIntent',
+					confirmationStatus: 'NONE',
+					slots: {
+						channelname: {
+							name: 'channelname',
+							value: sessionAttributes.channelName,
+							confirmationStatus: 'NONE',
+						},
+						longmessage: {
+							name: 'longmessage',
+							confirmationStatus: 'NONE',
+						},
 					},
-					longmessage: {
-						name: 'longmessage',
-						confirmationStatus: 'NONE',
-					},
-				},
-			})
-			.getResponse();
+				})
+				.getResponse();
+		}
+
+
 	},
 };
