+++ lambda/custom/resources/en-US.json
@@ -39,6 +39,7 @@
         "SIMILAR_USERS": "There are multiple users with similar username. Please tell the serial number of desired user. {usernames}.",
         "CONFIRM_MORE": "Okay, go on.",
         "CONFIRM_MORE_REPROMPT": "Please continue your message.",
+        "POST_LONG_MESSAGE_CONFIRMATION": "You want to send, {longmessage}, to channel {channelName}, is that correct?",
         "AUTH_ERROR": "Sorry you are currently signed out , please use the companion app to authenticate ",
         "CARD_TITLE": "Post A Message",
         "ERROR": "Sorry, I couldn't send your message right now.",
